apps/py/pyproject.toml
[project]
name = "toolkit-py"
version = "0.1.0"
description = "Python sidecar: conversion, chunking, sanitization, structured extraction backends for toolkit"
requires-python = ">=3.12"
# Framework deps only. Engine deps live in optional-dependencies groups so
# operators / CI can install just what they need:
#   pip install .[convert]          → only conversion is wired up
#   pip install .[all,dev]          → everything, plus test tooling
dependencies = [
  "fastapi>=0.115",
  "uvicorn[standard]>=0.32",
  "python-multipart>=0.0.20",
  "pydantic>=2.9",
]

[project.optional-dependencies]
convert = [
  "markitdown[all]==0.1.5",
  "docling>=2.14",
]
chunk = [
  "sentence-transformers>=3.3",
  "langchain-text-splitters>=0.3",
  "langchain-experimental>=0.3",
  "langchain-community>=0.3",
  "einops>=0.8",
  "torch>=2.5",
]
sanitize = [
  "presidio-analyzer>=2.2",
  "presidio-anonymizer>=2.2",
  "spacy>=3.8",
]
extract = [
  "instructor>=1.7",
  "openai>=1.58",
]
all = [
  "toolkit-py[convert,chunk,sanitize,extract]",
]
dev = [
  "pytest>=8.3",
  "pytest-asyncio>=0.24",
  "httpx>=0.28",
  "respx>=0.22",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/toolkit_py"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
# Unset so markers don't error when a test uses a category without it installed.
filterwarnings = [
  "ignore::DeprecationWarning",
]
